Just need styling
Make the question type actual buttons. So they will have a green bg and white text. Also for the difficulty make them buttons but the easy should be green w white text. Medium should be yellow with white text and hard should be red with white text. And change the output "How does y work" to color white. And remove all the <style> tags and add them into the env.